+++ src/python_version.rs
@@ -15,7 +15,11 @@ pub(crate) const DEFAULT_PYTHON_VERSION: RequestedPythonVersion = RequestedPytho
 };
 pub(crate) const DEFAULT_PYTHON_FULL_VERSION: PythonVersion = LATEST_PYTHON_3_13;
 
-pub(crate) const LATEST_PYTHON_3_8: PythonVersion = PythonVersion::new(3, 8, 20);
+pub(crate) const OLDEST_SUPPORTED_PYTHON_3_MINOR_VERSION: u16 = 9;
+pub(crate) const NEWEST_SUPPORTED_PYTHON_3_MINOR_VERSION: u16 = 13;
+pub(crate) const NEXT_UNRELEASED_PYTHON_3_MINOR_VERSION: u16 =
+    NEWEST_SUPPORTED_PYTHON_3_MINOR_VERSION + 1;
+
 pub(crate) const LATEST_PYTHON_3_9: PythonVersion = PythonVersion::new(3, 9, 21);
 pub(crate) const LATEST_PYTHON_3_10: PythonVersion = PythonVersion::new(3, 10, 16);
 pub(crate) const LATEST_PYTHON_3_11: PythonVersion = PythonVersion::new(3, 11, 11);
@@ -156,18 +160,17 @@ pub(crate) fn resolve_python_version(
     } = requested_python_version;
 
     match (major, minor, patch) {
-        (..3, _, _) | (3, ..8, _) => Err(ResolvePythonVersionError::EolVersion(
-            requested_python_version.clone(),
-        )),
-        (3, 8, None) => Ok(LATEST_PYTHON_3_8),
+        (..3, _, _) | (3, ..OLDEST_SUPPORTED_PYTHON_3_MINOR_VERSION, _) => Err(
+            ResolvePythonVersionError::EolVersion(requested_python_version.clone()),
+        ),
+        (3, NEXT_UNRELEASED_PYTHON_3_MINOR_VERSION.., _) | (4.., _, _) => Err(
+            ResolvePythonVersionError::UnknownVersion(requested_python_version.clone()),
+        ),
         (3, 9, None) => Ok(LATEST_PYTHON_3_9),
         (3, 10, None) => Ok(LATEST_PYTHON_3_10),
         (3, 11, None) => Ok(LATEST_PYTHON_3_11),
         (3, 12, None) => Ok(LATEST_PYTHON_3_12),
         (3, 13, None) => Ok(LATEST_PYTHON_3_13),
-        (3, 14.., _) | (4.., _, _) => Err(ResolvePythonVersionError::UnknownVersion(
-            requested_python_version.clone(),
-        )),
         (major, minor, Some(patch)) => Ok(PythonVersion::new(major, minor, patch)),
     }
 }
@@ -182,9 +185,6 @@ pub(crate) enum ResolvePythonVersionError {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    const OLDEST_SUPPORTED_PYTHON_3_MINOR_VERSION: u16 = 8;
-    const NEWEST_SUPPORTED_PYTHON_3_MINOR_VERSION: u16 = 13;
 
     #[test]
     fn python_version_url() {
@@ -239,10 +239,10 @@ mod tests {
     #[test]
     fn read_requested_python_version_python_version_file() {
         assert_eq!(
-            read_requested_python_version(Path::new("tests/fixtures/python_3.7")).unwrap(),
+            read_requested_python_version(Path::new("tests/fixtures/python_3.9")).unwrap(),
             RequestedPythonVersion {
                 major: 3,
-                minor: 7,
+                minor: 9,
                 patch: None,
                 origin: PythonVersionOrigin::PythonVersionFile,
             }
@@ -357,7 +357,7 @@ mod tests {
     fn resolve_python_version_unsupported() {
         let requested_python_version = RequestedPythonVersion {
             major: 3,
-            minor: NEWEST_SUPPORTED_PYTHON_3_MINOR_VERSION + 1,
+            minor: NEXT_UNRELEASED_PYTHON_3_MINOR_VERSION,
             patch: None,
             origin: PythonVersionOrigin::PythonVersionFile,
         };
@@ -370,7 +370,7 @@ mod tests {
 
         let requested_python_version = RequestedPythonVersion {
             major: 3,
-            minor: NEWEST_SUPPORTED_PYTHON_3_MINOR_VERSION + 1,
+            minor: NEXT_UNRELEASED_PYTHON_3_MINOR_VERSION,
             patch: Some(0),
             origin: PythonVersionOrigin::PythonVersionFile,
         };